+++ src/pharmpy/modeling/__init__.py
@@ -28,6 +28,7 @@ from pharmpy.modeling.covariate_effect import add_covariate_effect
 from pharmpy.modeling.data import (
     get_baselines,
     get_covariate_baselines,
+    get_doseid,
     get_doses,
     get_ids,
     get_mdv,
@@ -132,6 +133,7 @@ __all__ = [
     'get_baselines',
     'get_covariate_baselines',
     'get_doses',
+    'get_doseid',
     'get_ids',
     'get_number_of_individuals',
     'get_number_of_observations',